\name{largeScaleModelR-package}
\alias{largeScaleModelR-package}
\alias{largeScaleModelR}
\docType{package}
\title{
\packageTitle{largeScaleModelR}
}
\description{
\packageDescription{largeScaleModelR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{largeScaleModelR}
\packageIndices{largeScaleModelR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{largeScaleModelR}

Maintainer: \packageMaintainer{largeScaleModelR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
